requirements.txt
django
django-rest-framework
psycopg2
requests
django-filter
pytest
pytest-django